Adjust hero properties for alignment and style
Desired alignment for the hero is a 50/50 split between the left and
right. To get that I added flex: 1 1 0; to hero-left/right and a small
gap to the hero container. I added object-fit to the img so it would
fill the right hero container, but not constantly be changing
dimensions.

Button styling was added aswell to make it look nice